Clarifying when this package is required. (#192)
Fixed broken link and clarifying when this package needs to be used.
There seems to be some confusion on what the cookie law is and when it should be used. In an ideal world, nobody would ever see a cookie warning because having a cookie warning, means you are being tracked.